ft-162757626 Added app.py simplify testing cases
