+++ API/run.py
@@ -2,7 +2,7 @@ import os
 
 config_var = os.getenv('APP_SETTINGS')
 
-from app import create_app
+from app.app import create_app
 app = create_app(config_var)
 
 if __name__ == "__main__":
